Final implementation of the testing of the sales analysis methods

+++ labs/lab-week-3-sales-analysis/src/main/java/me/ronin/analysis/SaleRecord.java
@@ -2,5 +2,5 @@ package me.ronin.analysis;
 
 import java.time.LocalDate;
 
-public record SaleRecord(int saleId, LocalDate saleDate, long total, String productName) {
+public record SaleRecord(int saleId, LocalDate saleDate, double total, String productName) {
 }
